+++ src/app/Finance/Sales/bank-customer-receipt/bank-customer-receipt.component.ts
@@ -68,7 +68,9 @@ export class BankCustomerReceiptComponent implements OnInit {
     });
   }
   onDelete(index: any) {
-    this._purchaseService
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService
       .deleteBankPayment(this.bankPaymentTable[index].id)
       .then(
         (res: any) => {
@@ -77,6 +79,10 @@ export class BankCustomerReceiptComponent implements OnInit {
         (err: any) => {
         }
       );
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
