+++ src/app/Finance/Sales/cash-receipt-customer/cash-receipt-customer.component.ts
@@ -69,7 +69,9 @@ export class CashReceiptCustomerComponent implements OnInit {
     });
   }
   onDelete(index: any) {
-    this._purchaseService
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService
       .deleteBankPayment(this.cashReceiptPaymentTable[index].id)
       .then(
         (res: any) => {
@@ -78,6 +80,10 @@ export class CashReceiptCustomerComponent implements OnInit {
         (err: any) => {
         }
       );
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
