+++ src/app/Finance/Sales/cash-receipt-petty-customer/cash-receipt-petty-customer.component.ts
@@ -64,13 +64,19 @@ export class CashReceiptPettyCustomerComponent implements OnInit {
     });
   }
   onDelete(index: any) {
-    this._purchaseService
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService
       .deleteBankPayment(this.cashReceiptPettyPaymentTable[index].id)
       .then((res: any) => {
       window.location.reload();
     },
     (err: any) => {
     })
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
