+++ src/app/Finance/Sales/sales-non-tax-invoices/sales-non-tax-invoices.component.ts
@@ -81,13 +81,19 @@ export class SalesNonTaxInvoicesComponent implements OnInit {
     });
   }
   onDelete(index: any) {
-    this._purchaseService
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService
       .deletePurchseSales(this.salesNonTaxTableData[index].id)
       .then((res: any) => {
       window.location.reload();
     },
     (err: any) => {
     })
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
