+++ src/app/Finance/Sales/sales-order/sales-order.component.ts
@@ -90,12 +90,18 @@ export class SalesOrderComponent implements OnInit {
       });
   }
   onDelete(index: any) {
-    this._purchaseService.deletePurchseOrder(this.salesTableData[index].orderId).then((res:any)=>{
-      console.log('deleted',res);
-      window.location.reload();
-    },
-    (err: any) => {
-    })
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService.deletePurchseOrder(this.salesTableData[index].orderId).then((res:any)=>{
+        console.log('deleted',res);
+        window.location.reload();
+      },
+      (err: any) => {
+      })
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
