+++ src/app/Finance/cash-payment/cash-payment.component.ts
@@ -68,11 +68,17 @@ export class CashPaymentComponent implements OnInit {
     });
   }
   onDelete(index: any) {
-    this._purchaseService.deleteBankPayment(this.cashPaymentTable[index].id).then((res:any)=>{
-      window.location.reload();
-    },
-    (err: any) => {
-    })
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService.deleteBankPayment(this.cashPaymentTable[index].id).then((res:any)=>{
+        window.location.reload();
+      },
+      (err: any) => {
+      })
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
