+++ src/app/Finance/purchase-return/purchase-return.component.ts
@@ -83,11 +83,17 @@ export class PurchaseReturnComponent implements OnInit {
     })
   }
   onDelete(index: any) {
-    this._purchaseService.deletePurchseReturn(this.purchaseReturnTable[index].returnId).then((res:any)=>{
-      window.location.reload();
-    },
-    (err: any) => {
-    })
+    var text = "Are you sure to delete?";
+    if (confirm(text) == true) {
+      this._purchaseService.deletePurchseReturn(this.purchaseReturnTable[index].returnId).then((res:any)=>{
+        window.location.reload();
+      },
+      (err: any) => {
+      })
+    }
+    else {
+      alert('You pressed cancel');
+    }
   }
   printReceipt(id: any) {
